libc: arm: Slight improvement to memchr. (MWisBest)
When this memchr was submitted to newlib years back, ldmia was changed to ldrd.
It was suggested that it would be faster on A15, without a hit to A8/A9;
in fact, my A9 testcase showed a measurable improvement with this change.

Change-Id: I21fce7514706661a220fea953f5016cdfd795ed7

source: https://github.com/OptiPop/bionic/commit/99375ff58012eb8c166039020b441cb83efa1df9